refactor: move EngineValue components to separated files
