+++ site/source/components/ChiffreAffairesActivitéMixte.tsx
@@ -11,7 +11,8 @@ import { situationSelector } from '@/store/selectors/simulationSelectors'
 import { catchDivideByZeroError } from '@/utils'
 
 import { ExplicableRule } from './conversation/Explicable'
-import { Condition, WhenApplicable } from './EngineValue'
+import { Condition } from './EngineValue/Condition'
+import { WhenApplicable } from './EngineValue/WhenApplicable'
 import { SimulationGoal } from './Simulation'
 import { FromTop } from './ui/animate'
 import { useEngine } from './utils/EngineContext'
