+++ site/source/components/EngineValue/types.ts
@@ -1,0 +1,20 @@
+import { DottedName } from 'modele-social'
+import Engine, { ASTNode, PublicodesExpression } from 'publicodes'
+import React from 'react'
+
+export type ValueProps<Names extends string> = {
+	expression: PublicodesExpression
+	unit?: string
+	engine?: Engine<Names>
+	displayedUnit?: string
+	precision?: number
+	documentationPath?: string
+	linkToRule?: boolean
+	flashOnChange?: boolean
+} & React.HTMLAttributes<HTMLSpanElement>
+
+export type ConditionProps = {
+	expression: PublicodesExpression | ASTNode
+	children: React.ReactNode
+	engine?: Engine<DottedName>
+}
