+++ site/source/components/company/Details.tsx
@@ -9,7 +9,7 @@ import { H4 } from '@/design-system/typography/heading'
 import { Body } from '@/design-system/typography/paragraphs'
 
 import SeeAnswersButton from '../conversation/SeeAnswersButton'
-import Value from '../EngineValue'
+import Value from '../EngineValue/Value'
 
 export function CompanyDetails({
 	showSituation = false,
