+++ site/source/components/simulationExplanation/IndépendantCotisationsForfaitaires.tsx
@@ -1,6 +1,5 @@
 import { Trans } from 'react-i18next'
 
-import Value from '@/components/EngineValue'
 import { FromBottom } from '@/components/ui/animate'
 import { useEngine } from '@/components/utils/EngineContext'
 import { Markdown } from '@/components/utils/markdown'
@@ -9,6 +8,8 @@ import { Button } from '@/design-system/buttons'
 import { Spacing } from '@/design-system/layout'
 import { H3 } from '@/design-system/typography/heading'
 import { Intro } from '@/design-system/typography/paragraphs'
+
+import Value from '../EngineValue/Value'
 
 export default function CotisationsForfaitaires() {
 	const rule = useEngine().getRule(
