+++ site/source/pages/assistants/declaration-charges-sociales-independant/components/RésultatSimple.tsx
@@ -2,7 +2,8 @@ import { utils } from 'publicodes'
 import { Trans, useTranslation } from 'react-i18next'
 import { styled } from 'styled-components'
 
-import Value, { Condition } from '@/components/EngineValue'
+import { Condition } from '@/components/EngineValue/Condition'
+import Value from '@/components/EngineValue/Value'
 import { FromTop } from '@/components/ui/animate'
 import { useEngine } from '@/components/utils/EngineContext'
 import { Markdown } from '@/components/utils/markdown'
