+++ site/source/pages/assistants/declaration-charges-sociales-independant/index.tsx
@@ -7,7 +7,8 @@ import { styled } from 'styled-components'
 
 import { TrackPage } from '@/components/ATInternetTracking'
 import RuleInput from '@/components/conversation/RuleInput'
-import { Condition, WhenAlreadyDefined } from '@/components/EngineValue'
+import { Condition } from '@/components/EngineValue/Condition'
+import { WhenAlreadyDefined } from '@/components/EngineValue/WhenAlreadyDefined'
 import PageHeader from '@/components/PageHeader'
 import { FromTop } from '@/components/ui/animate'
 import Warning from '@/components/ui/WarningBlock'
