+++ site/source/pages/simulateurs/comparaison-statuts/components/DetailsRowCards.tsx
@@ -4,7 +4,9 @@ import { ReactNode } from 'react'
 import { Trans } from 'react-i18next'
 import { styled } from 'styled-components'
 
-import Value, { Condition, WhenNotApplicable } from '@/components/EngineValue'
+import { Condition } from '@/components/EngineValue/Condition'
+import Value from '@/components/EngineValue/Value'
+import { WhenNotApplicable } from '@/components/EngineValue/WhenNotApplicable'
 import RuleLink from '@/components/RuleLink'
 import { HelpIcon } from '@/design-system/icons'
 import { Grid } from '@/design-system/layout'
