+++ site/source/pages/simulateurs/comparaison-statuts/components/Détails.tsx
@@ -2,7 +2,8 @@ import { Trans, useTranslation } from 'react-i18next'
 import { styled } from 'styled-components'
 
 import { ExplicableRule } from '@/components/conversation/Explicable'
-import Value, { Condition } from '@/components/EngineValue'
+import { Condition } from '@/components/EngineValue/Condition'
+import Value from '@/components/EngineValue/Value'
 import { Accordion, Item } from '@/design-system'
 import { HelpButtonWithPopover } from '@/design-system/buttons'
 import { Emoji } from '@/design-system/emoji'
